+++ modules/acl/ui/src/sap/commerce/toolset/acl/editor/AclSplitEditor.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2026 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -21,64 +21,32 @@ package sap.commerce.toolset.acl.editor
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.actionSystem.PlatformDataKeys
 import com.intellij.openapi.application.edtWriteAction
-import com.intellij.openapi.fileEditor.FileEditorState
 import com.intellij.openapi.fileEditor.TextEditor
 import com.intellij.openapi.project.Project
-import com.intellij.openapi.util.Disposer
-import com.intellij.openapi.util.Key
-import com.intellij.openapi.util.UserDataHolderBase
-import com.intellij.openapi.util.getOrCreateUserData
-import com.intellij.openapi.vfs.VirtualFile
-import com.intellij.pom.Navigatable
-import com.intellij.ui.OnePixelSplitter
 import com.intellij.util.asSafely
 import kotlinx.coroutines.launch
 import sap.commerce.toolset.exec.context.DefaultExecResult
 import sap.commerce.toolset.impex.exec.context.ImpExExecContext
-import java.awt.BorderLayout
-import java.beans.PropertyChangeListener
+import sap.commerce.toolset.ui.editor.SplitEditorBase
 import java.io.Serial
-import javax.swing.JComponent
-import javax.swing.JPanel
 
 fun AnActionEvent.aclSplitEditor() = this.getData(PlatformDataKeys.FILE_EDITOR)
     ?.asSafely<AclSplitEditorEx>()
 
-class AclSplitEditorEx(override val textEditor: TextEditor, private val project: Project) : UserDataHolderBase(), AclSplitEditor {
+class AclSplitEditorBase(textEditor: TextEditor, project: Project) : SplitEditorBase(textEditor, project), AclSplitEditorEx {
 
     companion object {
         @Serial
         private const val serialVersionUID: Long = -3770445176190649196L
-
-        private val KEY_IN_EDITOR_RESULTS = Key.create<Boolean>("acl.in_editor_results.key")
     }
 
-    override var inEditorResults: Boolean
-        get() = getOrCreateUserData(KEY_IN_EDITOR_RESULTS) { true }
-        set(state) {
-            putUserData(KEY_IN_EDITOR_RESULTS, state)
-            verticalSplitter.secondComponent?.isVisible = state
-        }
+    override val inEditorResultsTitle = "ACL Execution Results"
 
-    private var inEditorResultsView: JComponent?
-        get() = verticalSplitter.secondComponent
-        set(view) {
-            verticalSplitter.secondComponent = view
-        }
-
-    private val verticalSplitter = OnePixelSplitter(true).apply {
-        isShowDividerControls = true
-        splitterProportionKey = "$javaClass.verticalSplitter"
-        setHonorComponentsMinimumSize(true)
-
-        firstComponent = textEditor.component
+    override fun showLoader(context: ImpExExecContext) {
+        inEditorResultsView = AclInEditorResultsView.getInstance(project).executingView(context.executionTitle)
     }
 
-    private val rootPanel = JPanel(BorderLayout()).apply {
-        add(verticalSplitter, BorderLayout.CENTER)
-    }
-
-    fun renderExecutionResult(execResult: DefaultExecResult) = AclInEditorResultsView.getInstance(project).resultView(this, execResult) { coroutineScope, view ->
+    override fun renderExecutionResult(result: DefaultExecResult) = AclInEditorResultsView.getInstance(project).resultView(this, result) { coroutineScope, view ->
         coroutineScope.launch {
             edtWriteAction {
                 inEditorResultsView = view
@@ -86,31 +54,5 @@ class AclSplitEditorEx(override val textEditor: TextEditor, private val project:
         }
     }
 
-    fun showLoader(context: ImpExExecContext) {
-        inEditorResultsView = AclInEditorResultsView.getInstance(project).executingView(context.executionTitle)
-    }
-
-    override fun addPropertyChangeListener(listener: PropertyChangeListener) {
-        textEditor.addPropertyChangeListener(listener)
-        component.addPropertyChangeListener(listener)
-    }
-
-    override fun removePropertyChangeListener(listener: PropertyChangeListener) {
-        textEditor.removePropertyChangeListener(listener)
-        component.removePropertyChangeListener(listener)
-    }
-
-    override fun getPreferredFocusedComponent(): JComponent? = verticalSplitter.firstComponent
-
-    override fun getComponent() = rootPanel
     override fun getName() = "Acl Split Editor"
-    override fun setState(state: FileEditorState) = textEditor.setState(state)
-    override fun isModified() = textEditor.isModified
-    override fun isValid() = textEditor.isValid && component.isValid
-    override fun dispose() = Disposer.dispose(textEditor)
-    override fun getEditor() = textEditor.editor
-    override fun canNavigateTo(navigatable: Navigatable) = textEditor.canNavigateTo(navigatable)
-    override fun navigateTo(navigatable: Navigatable) = textEditor.navigateTo(navigatable)
-    override fun getFile(): VirtualFile? = editor.virtualFile
-
 }