+++ modules/flexibleSearch/ui/src/sap/commerce/toolset/flexibleSearch/editor/FlexibleSearchSplitEditorEx.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2026 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -18,11 +18,16 @@
 
 package sap.commerce.toolset.flexibleSearch.editor
 
-import com.intellij.openapi.Disposable
-import sap.commerce.toolset.editor.ParameterizedSplitEditor
+import sap.commerce.toolset.flexibleSearch.exec.context.FlexibleSearchExecContext
+import sap.commerce.toolset.flexibleSearch.exec.context.FlexibleSearchExecResult
+import sap.commerce.toolset.ui.editor.SplitEditorEx
+import kotlin.time.Duration
+import kotlin.time.Duration.Companion.milliseconds
 
-interface FlexibleSearchSplitEditor : ParameterizedSplitEditor {
-    var csvResultsDisposable: Disposable?
-    var virtualParameters: Map<String, FlexibleSearchVirtualParameter>?
-    val virtualText: String
+interface FlexibleSearchSplitEditorEx : FlexibleSearchSplitEditor, SplitEditorEx {
+    var lastExecResult: FlexibleSearchExecResult?
+    fun showLoader(context: FlexibleSearchExecContext)
+    fun renderExecutionResult(result: FlexibleSearchExecResult)
+    fun clearExecutionResult()
+    fun refreshParameters(delayMs: Duration = 500.milliseconds)
 }