+++ modules/polyglotQuery/ui/src/sap/commerce/toolset/polyglotQuery/editor/PolyglotQuerySplitEditorEx.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for IntelliJ IDEA.
- * Copyright (C) 2019-2025 EPAM Systems <hybrisideaplugin@epam.com> and contributors
+ * Copyright (C) 2019-2026 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -16,13 +16,15 @@
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
 
-package sap.commerce.toolset.flexibleSearch.editor
+package sap.commerce.toolset.polyglotQuery.editor
 
-import com.intellij.openapi.Disposable
-import sap.commerce.toolset.editor.ParameterizedSplitEditor
+import sap.commerce.toolset.flexibleSearch.exec.context.FlexibleSearchExecResult
+import sap.commerce.toolset.ui.editor.SplitEditorEx
+import kotlin.time.Duration
+import kotlin.time.Duration.Companion.milliseconds
 
-interface FlexibleSearchSplitEditor : ParameterizedSplitEditor {
-    var csvResultsDisposable: Disposable?
-    var virtualParameters: Map<String, FlexibleSearchVirtualParameter>?
-    val virtualText: String
+interface PolyglotQuerySplitEditorEx : PolyglotQuerySplitEditor, SplitEditorEx {
+    fun showLoader(richMessage: String)
+    fun renderExecutionResult(result: FlexibleSearchExecResult)
+    fun refreshParameters(delayMs: Duration = 500.milliseconds)
 }